Add a minimum CLTV delta to prevent routing failures from LND's bad check
LND checks the wrong link's forwarding policy when determining whether an
HTLC has sufficient time lock to be forwarded.
(See: https://github.com/lightningnetwork/lnd/issues/1431)

To compensate for (but not fix) this issue, we use the default time lock
delta as a minimum. Since this routing is unlikely to make it into production,
we consider it good enough for now, until we remove this routing or LND
fixes the issue.