Add minimum timelock to exit hop to avoid LND's forwarding check
LND checks the inbound link's forwarding policy when determining
whether the HTLC extended to a final node has a valid time lock,
instead of using the payment request or global default, as
specified. (See: https://github.com/lightningnetwork/lnd/issues/1431#issuecomment-399323539)

To avoid as many issues as possible (but not actually fix it)
we use a minimum final time lock of 144.