Prepare for swap by creating an external preimage invoice on the engine
This change adds rHash and externalPreimage as options for the addInvoice
endpoint on LND, and uses those new parameters in the prepareSwap action.

The prepareSwap action is called on the Maker's inbound engine to mark
an invoice that will translate to the next chain.

This change also adjusts the memo field on "createSwapHash" so as to
allow for multiple "kinesis" memos.

Finally, this change uses the k#epic/cross-chain-preimage branch of
kinesis-exchange/lnd which supports these new LND features.